tcg: support exec memory locking
To support iOS which does not allow allocating RWX pages, we can only mmap
initially as RW and later mprotect it as RX. The full details are
documented in docs/devel/ios.rst

This largely reverts 6e3b2bfd6af488a896f7936e99ef160f8f37e6f2

On iOS, because we must lock/unlock executable memory for writing, it does
not make sense to store the TBs in executable memory. Instead, we will use
the beginning of each tcg region for TBs.

In order to make the most effective use of space, after a TLB flush, we
find the average size of a code block in previous allocations and then set
aside enough space for future allocations.

This is less space efficient than storing the TB before the code block but
can be more space efficient than the static allocations before that
implementation.

With multiple TCG contexts there is an issue of one vCPU executing a TB
generated by another vCPU and therefore the executable page unlocking can
happen during the execution.

For each TB, we keep a pointer to the tcg_ctx.code_locked_top_page and can
check that a given TB is executable if either 1) it was generated in the
same host thread as the tcg_ctx (then we would be running it after locking
the memory) or 2) it is below the water level for locked exec page (for
that TCG context). We will only return an existing TB in a search if it
matches either condition.

We do this to prevent having to (mutex) lock while generating code and
ruin the advantages of having multiple TCG threads. However, there is now
the possibility of two vCPUs generating the same TB and we have to just
deal with that overhead.