tcg: make mirror mapped JIT optional
On platforms that support mirror mapping, give the user the option to
disable it. There is a performance penalty for mirror JIT and on some
platforms, system security must be downgraded to use QEMU with mirror JIT
turned off.

To aid in testing, we still enable mirror JIT by default when configured
with --enable-debug-tcg.